# api-docs/swagger.yaml
basePath: /cosmos
definitions:
...
  utils.APIResponse:
    properties:
      data: {}
      status:
        example: OK
        type: string
    type: object
...
  utils.HTTPErrorResult:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for Cosmos Cloud server management
  title: Cosmos Server API
  version: 0.22.10
paths:
...
  /api/dns:
    get:
      description: Performs a DNS lookup for the given URL and returns the resolved
        IPv4 address
      parameters:
      - description: URL to resolve
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  type: string
              type: object
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      summary: Resolve DNS for a URL
      tags:
      - system
  /api/dns-check:
    get:
      description: Verifies that DNS is correctly configured for the given URL
      parameters:
      - description: URL to check DNS for
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Check DNS configuration for a URL
      tags:
      - system
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"